--- a/.gitignore
+++ b/.gitignore
@@ -12,5 +12,8 @@ __pycache__/
 .env
 .venv
 
+# VS Code
+.vscode/
+
 # Documentation
 src/mrg23k3a/docs/_build/