--- a/docs/conf.py
+++ b/docs/conf.py
@@ -19,17 +19,18 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../src'))
+
+sys.path.insert(0, os.path.abspath("../src"))
 
 
 # -- Project information -----------------------------------------------------
 
-project = 'mrg32k3a'
-copyright = '2022, David Eckman, Shane Henderson, and Sara Shashaani'
-author = 'David Eckman, Shane Henderson, and Sara Shashaani'
+project = "mrg32k3a"
+copyright = "2022, David Eckman, Shane Henderson, and Sara Shashaani"
+author = "David Eckman, Shane Henderson, and Sara Shashaani"
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.1'
+release = "0.0.1"
 
 
 # -- General configuration ---------------------------------------------------
@@ -38,27 +39,23 @@ release = '0.0.1'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.viewcode',
-    'sphinx.ext.coverage'
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.coverage",
 ]
 
-intersphinx_mapping = {'python': ('https://docs.python.org/3', None)}
+intersphinx_mapping = {"python": ("https://docs.python.org/3", None)}
 
 
-autodoc_mock_imports = ['numpy',
-                        'math',
-                        'random',
-                        'copy'
-]
+autodoc_mock_imports = ["numpy", "math", "random", "copy"]
 
 napolean_numpy_docstring = True
 
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -71,16 +68,17 @@ exclude_patterns = []
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'default'
+html_theme = "default"
 
-latex_engine = 'xelatex'
+latex_engine = "xelatex"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 
-#html_static_path = ['_static']
+# html_static_path = ['_static']
 html_static_path = []
 
 from sphinx.ext.apidoc import main
-main(["-o",  os.path.abspath('.'), os.path.abspath('../src'), "-f"])
+
+main(["-o", os.path.abspath("."), os.path.abspath("../src"), "-f"])