… .github/workflows/python-tests.yml
… .readthedocs.yaml
… README.md
… docs/index.rst
… docs/modules.rst
… docs/mrg32k3a.rst
… environment.yml
… pyproject.toml
… tests/test_matmodops.py