## Overview

This PR overhauls the `mrg32k3a` random number generator to significantly improve performance, reduce redundancy, and modernize the internal structure. Highlights:

- 30–40% faster distribution sampling and RNG calls  
- ~95% test coverage across all streams and variate generators  
- Removed outdated components and streamlined internal logic
- Added GitHub actions to ensure compatibility and consistency between Python and Numpy versions

---

## Key Changes

### `mrg32k3a`

- Replaced math-heavy logic with faster, minimal-overhead equivalents using Python's built-in types and operators
- Precomputed transition matrices (`A[1|2]p[47|94|141]`) and constants (`mrgm1_plus_1`, etc.) to eliminate redundant computation
- Internal stream states now use `np.ndarray` for efficient slicing and modular matrix math, while `seed()` still accepts a Python `tuple`
- Centralized and deduplicated modular state transitions via `_advance_state()` and `_apply_transition()` helpers
- Simplified stream/substream/subsubstream advancement logic
- Ensured consistency between return values and type hints throughout the codebase

---

## Testing

- Expanded test suite to include:
  - Stream reset and advancement behavior
  - Random variate correctness and distribution validation
  - Vector generation summation and bounds checking
- Tests reorganized by class for better structure and readability
- Achieved ~95% line coverage of generator logic
- Added GitHub actions to automatically run the testing suite with all supported versions of Python (minus 3.5 which had dependency issues)
- Scheduled the compatibility tests to run monthly to ensure timely catching of any issues with dependencies resulting in different results

---

## Performance Summary

| Component                             | Change       |
|--------------------------------------|--------------|
| RNG core (`random()`)                | **+32%**     |
| BSM (normal quantile approx)         | **+16%**     |
| Normal Variate Generation            | **+18%**     |
| Lognormal Variate Generation         | **+17%**     |
| Poisson Variate Generation           | **+22%**     |
| Gumbel Variate Generation            | **+39%**     |
| Binomial Variate Generation          | **+57%**     |
| Integer random vector (simplex)      | **+16%**     |
| Continuous random vector             | **+13–14%**  |
| Stream advancement (`start_*`)       | **+7–9%**     |
| Reset methods (`reset_*`)            | **−93% to −110%**<br><sub>(See note below)</sub> |
| Start Fixed `S_SS_SSS`               | **+6–17%**   |

> **Note on regressions:**  
> The `reset_*()` methods show a microbenchmark-level regression due to the addition of a `tuple()` conversion from NumPy arrays. This cost is on the order of microseconds, and resets are typically infrequent in most applications. Overall performance remains excellent even in scenarios with frequent resets.

---

## Summary

This refactor delivers a leaner, faster, and more maintainable implementation of `mrg32k3a`, with significant speedups in all key random variate generation paths and much broader test coverage. Ready for review and merge.