@Grochocinski 2025-03-19
## Speed Up Testing Results (TLDR ~33% Improvement)
### Generating 1M Random Values
Old Avg: 1.88 seconds
New Avg: 2.79 seconds
Improvement: **32.54**%

### Generating 10M Random Values
Old Avg: 34.48 seconds
New Avg: 51.54 seconds
Improvement: **33.10**%

*sample size of 10 runs each for old and new executed in randomized order, tested on Windows 11 laptop w/ Snapdragon X Plus processor*