@Grochocinski 2025-03-24
Alright, I think it's actually good to go at this point (passes all tests / Ruff formatting/linting checks / type checks, as well as showing substantial performance uplift and having easy-to-read code)